% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbinom.R
\name{gbinom}
\alias{gbinom}
\title{Graph of the Binomial Distribution with point and region probabilities
illustrated.}
\usage{
gbinom(n, p, low = 0, high = n, scale = F, a = NA, b = NA, findquantile = NA)
}
\arguments{
\item{n}{Number of trials.}

\item{p}{Probability of a Success.}

\item{low}{Lower limit of the X axis scale.  Set to zero by default.}

\item{high}{Upper limit of the X axis scale.  Set to N, the number of
trials, by default.}

\item{scale}{Truncates the scale when TRUE.  Shows full scale when FALSE}

\item{a}{Arguments \code{a} and/or \code{b} can be used to set regions of
the RV scale for which to color code and compute probabilities.  See the
examples below for several illustrations.  The \code{a} value would
typically be the smaller value if both \code{a} and \code{b} are used.}

\item{b}{Arguments \code{a} and/or \code{b} can be used to set regions of
the RV scale for which to color code and compute probabilities.  See the
examples below for several illustrations.  The \code{a} value would
typically be the smaller value if both \code{a} and \code{b} are used.}

\item{findquantile}{Takes a cumulative left tail probability for
which a quantile is returned. Default is NA.  Arguments \code{a}
and \code{b} override \code{findquantile}}
}
\description{
Binomial Distributions of an N and p values can be plotted.  Scale limits
can be specified, as well as regions of the RV scale for determining
probabilities.  These regions are indicated with color on the graph and
probabilities are also displayed.
}
\examples{

gbinom(4,.5)
gbinom(100,.5)
gbinom(4,.4)
gbinom(10,.4)
gbinom(25,.7)
gbinom(4,.5,a=3)
gbinom(4,.5,b=1)
gbinom(4,.5,a=3,b=1)
gbinom(100,.667,a=60,b=70)
# improve the previous graph by resetting the x axis limits
gbinom(100,.667,a=60,b=70,low=45,high=85)

}
\references{
This function is a modification of a function originally coded
by Brett Larget, Univ. Wisconsin.\cr
http://www.stat.wisc.edu/~larget/R/prob.R
}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
