% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggeom.R
\name{ggeom}
\alias{ggeom}
\title{Graph of Probabilities of Geometric Distributions.}
\usage{
ggeom(k, p)
}
\arguments{
\item{k}{The number of trials to display on the X axis.}

\item{p}{Probability of a success}
}
\description{
Graph the probabilities of the first success on the rth trial where the
elementary events are sampled from a stationary bernoulli process.
}
\details{
The graph plots probabilities of 0:k failures before the first success.
}
\note{
This function is a modification of code from the Rcmdr package.
}
\examples{

ggeom(2,.5)
ggeom(10,.5)
ggeom(10,.2)
ggeom(15,.7)

}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
