% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnbinom.R
\name{gnbinom}
\alias{gnbinom}
\title{Graphical Depiction of the Negative Binomial Distribution}
\usage{
gnbinom(k, r, p)
}
\arguments{
\item{k}{this argument is used to set the x axis scale of the graph.  Often
choosing a value of 10 or 15 is adequate for textbook types of
demonstrations.  See the examples below for better illustration.}

\item{r}{The number of successes}

\item{p}{probability of a success on each trial.}
}
\description{
\code{gbinom} permits specification of regions in a negative binomial
distribution
}
\details{
The function plots probabilities of each of several possible numbers of
failures (ranging up to k) before the rth success.  The graph can be
confusing without exploring the examples below (as well as class materials
on the function).
}
\note{
This function is a modification of the code for the Negative Binomial
Distribution found in the \code{Rcmdr} package.
}
\examples{

#setting \code{r} to 1 reduces the negative binomial (Pascal) distribution
#to a geometric distribution and the plot can be easily interpreted.
gnbinom(5,1,.5)
# probabilities of 0,1,2,3 or 4 failures before the second success.
gnbinom(4,2,.5)
# probabilities of 0 through 8 failures before the second success.
gnbinom(8,2,.5)
# probabilities of 0 through 10 failures before the third success.
gnbinom(10,3,.5)
# any value of p can be used
gnbinom(10,2,.1)

}
\seealso{
\code{\link[bcdstats:gbinom]{gbinom}}, for an analogous function
displaying the binomial distribution.
}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
