% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnorm.R
\name{gnorm}
\alias{gnorm}
\title{Graphing a Normal Distribution (from Brett Larget)}
\usage{
gnorm(mu, sigma, a = NA, b = NA, findquantile = NA)
}
\arguments{
\item{mu}{\code{mu} is the Normal Distribution mean. Required.  Use 0 for
std normal.}

\item{sigma}{\code{sigma} is the Normal Distribution Std Deviation.
Required.  Use 1.0 for std normal.}

\item{a}{optional specification of a quantile value that can serve as a
"cutpoint".  \code{gnorm} will shade the tail region demarcated by this
value.  \code{a} can be used in conjunction with \code{b} to find
probabilities between the two values.}

\item{b}{optional specification of a quantile value that can serve as a
"cutpoint".  \code{gnorm} will shade the tail region demarcated by this
value.  \code{b} can be used in conjunction with \code{a} to find
probabilities between the two values.}

\item{findquantile}{Takes a cumulative left tail probability for
which a quantile is returned. Default is NA.  Arguments \code{a}
and \code{b} override \code{findquantile}}
}
\description{
\code{gnorm} plots a density function for a normal distribution of specified
mean and standard deviation.  It will also accept specification of one, or
two "cutpoints" and shade the corresponding tail regions or the region
between two cutpoints.  The function also displays the corresponding tail or
region probability.
}
\note{
This function can be very helpful in an instructional setting.
}
\examples{

gnorm(0,1)
gnorm(0,1,a=1.65)
gnorm(0,1,a=-1.65)
gnorm(0,1,b=1.65)
gnorm(0,1,a=1.65,b=1.65)
gnorm(0,1,a=1.96,b=-1.65)
gnorm(100,15,a=130)
gnorm(100,15,a=85)
gnorm(100,15,a=85,b=115)
gnorm(800,20,a=837,b=763)

}
\references{
This function is a modification of a function coded by Brett
Larget, Univ Wisconsin.\cr http://www.stat.wisc.edu/~larget/R/prob.R
}
\seealso{
\code{\link[bcdstats:pgnorm]{pgnorm}},
\code{\link[bcdstats:pgnormz]{pgnormz}},
\code{\link[bcdstats:qgnorm]{qgnorm}},
\code{\link[bcdstats:qgnormz]{qgnormz}}
}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
