% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howell2_1t.R
\docType{data}
\name{howell2_1t}
\alias{howell2_1t}
\title{Table 2.1 Data from Howell text}
\format{
A data frame with 300 observations on the following 4 variables.
\describe{ \item{trial}{a numeric vector} \item{rxtime}{a
numeric vector} \item{nstim}{a numeric vector}
\item{yesno}{a numeric vector} }
}
\description{
Howell's table 2.1
}
\examples{
data(howell2_1t)
## maybe str(howell2_1t) ; plot(howell2_1t) ...

}
\keyword{datasets}
