% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrinfo.R
\name{mrinfo}
\alias{mrinfo}
\title{Regression Information for lm class objects}
\usage{
mrinfo(lm.fit, cilevel = 0.95, minimal = TRUE)
}
\arguments{
\item{lm.fit}{A \code{lm} class model object.}

\item{cilevel}{Confidence interval level.  Default is .95.}

\item{minimal}{specifies volume of returned information (see details). Default=T.}
}
\description{
The \code{mrinfo} function produces a wealth of additional information
on multiple regression models above and beyond what summary(), anova(),
and Anova() produce.
}
\details{
The function takes an \code{lm} fit object.  Supplemental information is
calculated for \code{lm} fit objects.  When \code{minimal=T} is specified,
a data frame of several indices (see values) is returned.  When \code{minimal=F},
a list of information, including the supplemental values is returned:
zero-order correlations, the coefficients table, confidence intervals for the
regression coefficients, type I SS anova table and type III SS Anova table.
}
\section{Values of the Supplemental Information}{

\tabular{ll}{
\code{beta wt} \tab Standardized Regression Coefficients\cr
\code{partial r} \tab Partial correlations of IVs with the DV\cr
\code{semi-partial r} \tab Semi-partial correlations of IVs with the DV\cr
\code{tolerances} \tab Tolerance for each IV\cr
\code{unique} \tab Unique proportion of variance in DV accounted for
by each IV\cr
\code{common}  \tab Common proportion of variance in DV accounted for
by each IV shared with other IVs\cr
\code{total} \tab Total proportion of variance in DV accounted for
by each IV\cr
\code{pearsons}  \tab The zero-order pearson correlation matrix among
all variables\cr
}
}

\section{Warnings}{


The \code{mrinfo} function is designed to work with multiple regression objects
where an intercept is estimated.  Models where the intercept is forced through
the origin ('no intercept' models) are problematic for interpretation of the
supplemental information listed above and it is not returned.

Simple regression models do not require the supplemental information, but the
user can specify \code{minimal=F} to obtain the longer list of detailed information.
}

\examples{
data(attitude)
fit1 <- lm(rating ~ complaints + learning +  privileges, data=attitude)
#summary(fit1)
mrinfo(fit1, minimal=TRUE, cilevel=.99)
mrinfo(fit1, minimal=FALSE, cilevel=.95)

}
\references{
This function is a modification of a function originally coded
in the \code{regr} function in the \bold{yhat} package.  It uses the
\code{effect.size} and \code{commonalityCoefficients} functions from that package.
}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
