% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgchisq.R
\name{pgchisq}
\alias{pgchisq}
\title{Graphical display for the Chi Squared distribution, given a specified
quantile}
\usage{
pgchisq(quantile, df = 1)
}
\arguments{
\item{quantile}{The Chi Squared value to be evaluated, (\emph{e.g., 3.84)}}

\item{df}{The \code{df} for the \emph{Chi Squared} Distribution.  Default is
1 df (see note below)}
}
\description{
A function to produce a graph of a Chi Squared distribution with regions
shaded according to a specified quantile ( Chi Squared value).
}
\note{
The scaling on plots of Chi Squared Distributions of varying degrees
of freedom requires considerable adjustment. When df=1, \code{pgchisq}
returns a two-paneled figure.  The plot on the left is a fairly full scale
plot of the distribution without shading of a tail region.  The plot on the
right has axes truncated so that scale expansion will permit visualization
of the relevant area under the curve.

Note that the \code{pgchisq} function only displays upper tail
regions since most inference with the Chi Squared distribution
only requires upper tail evaluation.  The lower tail probability can be
found as 1 minus the returned upper tail probability.
}
\section{Warning}{
 Specifying extremely high or low quantiles may not yield
visible shaded areas if the regions are beyond the limits of the X axis
scale of the graph.
}

\examples{

pgchisq(3.84,df=1)
pgchisq(4.20,df=2)
pgchisq(30.5,df=25)
pgchisq(200,df=250)
pgchisq(4.75,df=1)
pgchisq(4.75,df=2)
pgchisq(45,df=25)
pgchisq(475,df=250)

}
\seealso{
\code{\link[bcdstats:qgchisq]{qgchisq}}
}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
