% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgf.R
\name{pgf}
\alias{pgf}
\title{Graphical display for the F distribution, given a specified quantile}
\usage{
pgf(quantile, df1 = 1, df2 = 5)
}
\arguments{
\item{quantile}{The quantile value (F value) to be evaluated.}

\item{df1}{Numerator degrees of freedom for the \emph{F} distribution.  Note
that the default is 1.}

\item{df2}{Denominator degrees of freedom for the \emph{F} distribution.
Note that the default is 25.}
}
\description{
A function to draw an F distribution and shade a region defined by the
quantile.  Note that pgf only gives upper tail probabilities and upper tail
shading of the relevant area under the curve.
}
\note{
This function can be very helpful in instructional situations and can
replace the "F table" from textbooks.

Warning: Specifying extremely high quantiles may not yield visible shaded
areas if the quantile is in a region where the distribution approaches a
density of zero. \cr In addition, probabilities are rounded to five decimal
places and will display as zero if below .000005.  Use the \code{pf}
function in this case.

Note that the \code{pgf} function only displays upper tail
regions since most inference with the F distribution
only requires upper tail evaluation.  The lower tail probability can be
found as 1 minus the returned upper tail probability.
}
\examples{
pgf(3.75, df1=1,df2=25)
pgf(3.75, df1=2,df2=50)
pgf(3.75, df1=1,df2=125)
pgf(3.75, df1=6,df2=80)
pgf(4.11, df1=1,df2=18)
pgf(.65, df1=1,df2=25)
}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
