% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgnormz.R
\name{pgnormz}
\alias{pgnormz}
\title{Graphical display for a Standard Normal distribution, given a specified
quantile ("Z") value}
\usage{
pgnormz(quantile, tail = "upper")
}
\arguments{
\item{quantile}{The quantile value ("Z" value) to be evaluated.  Note that
quantile should be positive if a two-tailed result is requested.}

\item{tail}{Area(s) to be shaded on the graph.  If \code{"upper"}, the
function returns a graph with the upper tail area shaded and the upper tail
probability displayed.  If \code{"lower"}, the function returns a graph with
the lower tail area shaded and the lower tail probability displayed.  If
\code{"two"} the function returns a graph with both tails shaded and a
summed two-tailed probability displayed.}
}
\description{
A function to draw a Standard Normal distribution and shade a region defined
by the quantile argument
}
\section{Warning}{
 Specifying extremely high or low quantiles may not yield
visible shaded areas if the regions are beyond the limits of the X axis
scale of the graph. \cr In addition, probabilities are rounded to five
decimal places and will display as zero if below .000005.  Use the
\code{pnorm} function in this case.
}

\examples{
#shows upper tail by default
pgnormz(1.2)
pgnormz(1.0,tail="upper")
pgnormz(-1.64,tail="lower")
pgnormz(2.0,tail="lower")
pgnormz(1.96,tail="upper")

}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
