% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgt.R
\name{pgt}
\alias{pgt}
\title{Graphical display for the "Student's" t distribution, given a specified
quantile}
\usage{
pgt(quantile, df = 100, tail = "upper")
}
\arguments{
\item{quantile}{The quantile value (t value) to be evaluated.  Note that
quantile should be positive if a two-tailed result is requested.}

\item{df}{Degrees of freedom for the t distribution.  Note that the default
is 100.}

\item{tail}{Area(s) to be shaded on the graph.  If \code{"upper"}, the
function returns a graph with the upper tail area shaded and the upper tail
probability displayed.  If \code{"lower"}, the function returns a graph with
the lower tail area shaded and the lower tail probability displayed.  If
\code{"two"} the function returns a graph with both tails shaded and a
summed two-tailed probability displayed.}
}
\description{
A function to draw a Student's t distribution and shade a region defined by
the quantile
}
\note{
This function can be very helpful in instructional situations and can
replace the "t table" from textbooks.
}
\section{Warning}{
 Specifying extremely high or low quantiles may not yield
visible shaded areas if the regions are beyond the limits of the X axis
scale of the graph. \cr In addition, probabilities are rounded to five
decimal places and will display as zero if below .000005.  Use the \code{pt}
function in this case.
}

\examples{
pgt(-1.76, df=25, tail="lower")
pgt(1.76, df=25, tail="upper")
pgt(1.76, df=25, tail="two")
pgt(2.31, df=10, tail="lower")
# don't use this next example form with tail="two"
# instead use the following example example form
pgt(-1.7, df=10, tail="two")
pgt(1.7, df=10, tail="two")

}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
