% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.corr.R
\name{print.adjust.corr}
\alias{print.adjust.corr}
\title{Print method for adjust.corr}
\usage{
\method{print}{adjust.corr}(x, ...)
}
\arguments{
\item{x}{An object produced by \code{adjust.corr}}

\item{...}{not used}
}
\description{
Print method for adjust.corr
}
\seealso{
Used with the \code{\link{adjust.corr}} function
}
