% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgchisq.R
\name{qgchisq}
\alias{qgchisq}
\title{Graphical display for the Chi Squared distribution, given a specified
probability}
\usage{
qgchisq(prob, df = 1)
}
\arguments{
\item{prob}{The tail probability value to define the region to be shaded on
the graph and the corresponding quantile to be displayed.}

\item{df}{The \code{df} for the \emph{Chi Squared} Distribution.  Default is
1 df (see note below)}
}
\description{
A function to produce a graph of a Chi Squared distribution with regions
shaded according to a specified probability.
}
\note{
The scaling on plots of Chi Squared Distributions of varying degrees
of freedom requires considerable adjustment. When df=1, \code{qgchisq}
returns a two-paneled figure.  The plot on the left is a fairly full scale
plot of the distribution without shading of a tail region.  The plot on the
right has axes truncated so that scale expansion will permit visualization
of the relevant area under the curve.

Also note that the \code{qgchisq} function only displays upper tail regions since
most inference with the Chi Squared distribution only requires upper tail
evaluation.  The lower tail probability is 1 minus the returned value.
}
\section{Warning}{
 Specifying extremely high or low probabilities may not
yield visible shaded areas if the regions are beyond the limits of the X
axis scale of the graph.
}

\examples{

qgchisq(.05,df=1)
qgchisq(.05,df=2)
qgchisq(.05,df=25)
qgchisq(.05,df=250)
qgchisq(.01,df=1)
qgchisq(.01,df=2)
qgchisq(.01,df=25)
qgchisq(.01,df=250)

}
\seealso{
\code{\link[bcdstats:pgchisq]{pgchisq}}
}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
