% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgf.R
\name{qgf}
\alias{qgf}
\title{Graphical display for the F distribution, given a specified probability}
\usage{
qgf(prob, df1 = 1, df2 = 5)
}
\arguments{
\item{prob}{The upper tail probability value to define the region to be
shaded on the graph and the corresponding quantile to be displayed.}

\item{df1}{Numberator degrees of freedom for the \emph{F} distribution.
Note that the default is 1.}

\item{df2}{Numberator degrees of freedom for the \emph{F} distribution.
Note that the default is 25.}
}
\description{
A function to produce a graph of an F distribution with regions shaded
according to a specified probability.
}
\note{
This function can be very helpful in instructional situations and can
replace the "F table" from textbooks.

Also note that the \code{qgf} function only displays upper tail regions since
most inference with the F distribution only requires upper tail
evaluation.  The lower tail probability is 1 minus the returned value.
}
\section{Warning}{
 Specifying extremely small probabilities may not yield
visible shaded areas if the quantile is in a region of the graph where the
density approaches zero.
}

\examples{

qgf(.05,df1=1,df2=25)
qgf(.05,df1=6,df2=80)
qgf(.35,df1=1,df2=25)
qgf(.01,df1=3,df2=32)

}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
