% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgnorm.R
\name{qgnorm}
\alias{qgnorm}
\title{Graphical display for a normal distribution, given a specified tail
probability}
\usage{
qgnorm(prob, mean = 0, sd = 1, tail = "upper")
}
\arguments{
\item{prob}{The tail probability value to define the region to be shaded on
the graph and the corresponding quantile to be displayed.  note that if
\code{tail="two"}, this probability is halved for display in each tail.}

\item{mean}{\code{mean} is the Distribution Mean}

\item{sd}{\code{sd} is the Distribution Standard Deviation}

\item{tail}{Area(s) to be shaded on the graph.  If \code{"upper"}, the
function returns a graph with the upper tail area shaded and the upper tail
probability displayed.  If \code{"lower"}, the function returns a graph with
the lower tail area shaded and the lower tail probability displayed.  If
\code{"two"} the function returns a graph with both tails shaded and the
specified probability is halved between these two tails. tail=\code{"upper"}
is the default.}
}
\description{
A function to draw a normal distribution and shade the region(s) defined by
the probability argument
}
\note{
This function can be very helpful in instructional situations and can
replace the "t table" from textbooks.
}
\section{Warning}{
 Specifying extremely high or low probabilities may not
yield visible shaded areas if the regions are beyond the limits of the X
axis scale of the graph.
}

\examples{

qgnorm(.35,mean=2000,sd=400) #shows upper tail by default
qgnorm(.16,mean=100,sd=15,tail="upper")
qgnorm(.16,mean=100,sd=15,tail="lower")
qgnorm(.975,mean=100,sd=15,tail="lower")
qgnorm(.025,mean=0,sd=1,tail="upper")
#note that pgnormz also yields the std normal
#without mean and sd arguments required#'
}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
