% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgnormz.R
\name{qgnormz}
\alias{qgnormz}
\title{Graphical display for a Standard Normal distribution, given a specified tail
probability}
\usage{
qgnormz(prob, tail = "upper")
}
\arguments{
\item{prob}{\code{prob} is the tail probability value to define the region
to be shaded on the graph and the corresponding quantile to be displayed.
note that if \code{tail="two"}, this probability is halved for display in
each tail.}

\item{tail}{Area(s) to be shaded on the graph.  If \code{"upper"}, the
function returns a graph with the upper tail area shaded and the upper tail
probability displayed.  If \code{"lower"}, the function returns a graph with
the lower tail area shaded and the lower tail probability displayed.  If
\code{"two"} the function returns a graph with both tails shaded and the
specified probability is halved between these two tails. tail=\code{"upper"}
is the default.}
}
\description{
A function to draw a Standard Normal distribution and shade the region(s)
defined by the probability argument
}
\note{
This function can be very helpful in instructional situations and can
replace the "Z table" from textbooks.
}
\section{Warning}{
 Specifying extremely high or low probabilities may not
yield visible shaded areas if the regions are beyond the limits of the X
axis scale of the graph.
}

\examples{

qgnormz(.35) #shows upper tail by default
qgnormz(.16,tail="upper")
qgnormz(.16,tail="lower")
qgnormz(.975,tail="lower")
qgnormz(.025,tail="upper")

}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
