% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgt.R
\name{qgt}
\alias{qgt}
\title{Graphical display for the "Student's" t distribution, given a specified
probability}
\usage{
qgt(prob, df = 1, tail = "upper")
}
\arguments{
\item{prob}{The tail probability value to define the region to be shaded on
the graph and the corresponding quantile to be displayed.}

\item{df}{Degrees of freedom for the \emph{t} distribution.  Note that the
default is 100.}

\item{tail}{Area(s) to be shaded on the graph.  If \code{"upper"}, the
function returns a graph with the upper tail area shaded and the upper tail
probability displayed.  If \code{"lower"}, the function returns a graph with
the lower tail area shaded and the lower tail probability displayed.  If
\code{"two"} the function returns a graph with both tails shaded and the
specified probability is halved between these two tails.}
}
\description{
A function to produce a graph of a "Student's" t distribution with regions
shaded according to a specified probability.
}
\note{
This function can be very helpful in instructional situations and can
replace the "t table" from textbooks.
}
\section{Warning}{
 Specifying extremely high or low probabilities may not
yield visible shaded areas if the regions are beyond the limits of the X
axis scale of the graph.
}

\examples{

qgt(.05,df=25,tail="upper")
qgt(.01, df=12,tail="lower")
qgt(.90, df=15, tail="lower")
qgt(.01,df=50, tail="two")

}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
