% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsm.R
\name{rsm}
\alias{rsm}
\title{Robust Symmetry Plot}
\usage{
rsm(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{\code{x} is the quantiative univariate RV of interest}

\item{alpha}{use to set confidence limits, default =.05}

\item{...}{Will accept some other graphical parameters}
}
\description{
A symmetry plot with simulated 95 per cent boundaries based on resampling
methods
}
\examples{

data(howell2_1t)
rsm(howell2_1t$rxtime) # a reaction time variable

}
\references{
\code{rsm} is based on a function provided by
\emph{http://zoonek2.free.fr/UNIX/48_R/03.html}
}
\author{
Bruce Dudek's modification of Vincent Zoonekynd's function
}
