% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshinyApp.R
\name{runshinyApp}
\alias{runshinyApp}
\title{runshinyApp, a function to run shiny apps found in bcdstats}
\usage{
runshinyApp(appname)
}
\arguments{
\item{appname}{The name of the shiny app (IN QUOTES)}
}
\description{
Several Shiny apps are available in this package.  They are useful
as instructional tools for visualizations in an interactive/dynamic
framework.  Use the app name as the argument in the \code{runshinyApp} function.
}
\section{Probability Distribution Apps}{
 These can be used to find probabilities
or to find quantiles.  All apps provide a visulization with appropriate regions of
the distribution shown.
\itemize{
\item
\strong{stdnormal}: The standard normal distribution
\item
\strong{binomial}:  The binomial distribution
\item
\strong{tdist}: The student's t distribution (central t)
\item
\strong{chisqdist}:  The Chi squared distribution
\item
\strong{fdist}:  The F distribution
}
}

\section{Univariate Descriptive Statistics and Graphical depiction (EDA)}{
 Examples of
univariate plot types and several data sets.
\itemize{
\item
\strong{describe}: Visualize several types of frequency histograms, boxplots, violinplots, etc
}
}

\section{Sums of Squares}{
 Visualize the Sums of Squares and Variance calculation
\itemize{
\item
\strong{vizualizess}: A geometric approach to understanding Sums of Squares and Variance
}
}

\section{Sampling Distribution Simulation}{
 Visualize sampling distributions of several
descriptive statistics using differing initial random variable distributions.
\itemize{
\item
\strong{sampdist}: Simulate sampling distributions of several descriptive statistics
}
}

\section{Confidence Interval Visualization}{
 Visualize confidence intervals, "in the long run"
\itemize{
\item
\strong{confidence}: Simulate confidence intervals based on either t or z distributions
}
}

\section{Type I and II error visualization}{
 Visualize null and alternative sampling distributions
of various characteristics and consequent Type I and II error rates.
\itemize{
\item
\strong{betaprob}: Simulate overlapping null/alternative sampling distributions to visualize Type I
and II error rates
}
}

\section{Correlation/Regression Simulation Apps}{
 Simulate bivariate data and visualize
the components of bivariate correlation/covariance and simple regression.
\itemize{
\item
\strong{rectangles}: Visualize the Covariance/SP components
\item
\strong{corrsim}:  Simulate bivariate correlation and simple regression.  Visualize yhats.
}
}

\section{Trend Analysis}{
 Visualize components of orthogonal polynomial trend
\itemize{
\item
\strong{trend}: Simulate application of orthogonal polynomial trend to a one-factor ANOVA design.
}
}

\section{Interaction and Moderation}{
 Visualize interactions, moderator effects, simple effects
\itemize{
\item
\strong{mod2}: In two-IV linear models (regression and ANOVA), visualize two-way interactions with
simple effects, simple slopes, and regression surfaces.
}
}

\examples{

\dontrun{
runshinyApp("stdnormal")
# to see the list of available apps
runshinyApp()
}
}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
