% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrtests.R
\name{test2r.ind}
\alias{test2r.ind}
\title{Test Two Correlations from Independent Groups}
\usage{
test2r.ind(r1, r2, n1, n2, twotailed = TRUE)
}
\arguments{
\item{r1}{The Pearson correlation coefficient (rxy) in group 1.}

\item{r2}{The Pearson correlation coefficient (rxy) in group 2.}

\item{n1}{Sample size in group 1}

\item{n2}{Sample size in group 1}

\item{twotailed}{The test can be either two- or one-tailed by specifying
\code{twotailed=T} or \code{twotailed=F}, respectively.}
}
\description{
Differences in Pearson correlations of two variables are tested when they
are measured in independent samples, for example r(xy) in group 1 vs r(xy)
in group 2.
The method employed is one that utilizes Fisher's Z transformation of the
Pearson correlation coefficients.  The test statistic is a standard normal
deviate.  The method can be found in standard textbook sources (\emph{e.g.},
Hays, 1993, Howell, 2013.)  If the user wishes to perform tests of two r's
from independent groups with standard methods that do no utilize Fisher's Z
transform, they are encouraged to use linear regression models that employ
an interaction term.
}
\section{Related Functions}{
  \code{test2r.ind} is a member of a set of
functions that provide tests of differences between independent and
dependent correlations.  See:
\itemize{
\item
\code{\link[bcdstats:test2r.t2]{test2r.t2}} Test two dependent correlations
with the the T2 method: r(yx1) vs r(yx2)
\item
\code{\link[bcdstats:test2r.mengz1]{test2r.mengz1}}, Test the difference between
two dependent correlations with the the Meng z1 method: r(yx1) vs r(yx2)in one
sample of cases.
\item
\code{\link[bcdstats:test2r.steigerz1]{test2r.steigerz1}}, Test the difference between
two dependent correlations with the the Steiger z1 method: r(yx1) vs r(yx2) in one
sample of cases.
\item
\code{\link[bcdstats:test2r.steigerz2]{test2r.steigerz2}}, Test the difference between
two dependent correlations with the the Steiger z2 method: r(jk) vs r(hm) in one
sample of cases.
\item
\code{\link[bcdstats:test2r.ind]{test2r.ind}}, the present function }
}

\examples{

test2r.ind(.30,.35,n1=50,n2=60)
test2r.ind(.10,.45,n1=60,n2=80)
test2r.ind(.41,.59,n1=100,n2=105)
test2r.ind(.41,.59,n1=100,n2=105,twotailed=FALSE)


}
\references{
Cheung, M. W. L., & Chan, W. (2004). Testing dependent
correlation coefficients via structural equation modeling.
\emph{Organizational Research Methods}, 7(2), 206-223. \cr Dunn, O. J., &
Clark, V. (1971). Comparison of tests of the equality of dependent
correlation coefficients. \emph{Journal of the American Statistical
Association}, 66(336), 904-908. \cr Hays, W. L. (1994). \emph{Statistics}
(5th ed.). Fort Worth: Harcourt College Publishers.\cr Hendrickson, G. F.,
Stanley, J. C., & Hills, J. R. (1970). Olkin's new formula for significance
of r13 vs. r23 compared with Hotelling's method. \emph{American Educational
Research Journal}, 7(2), 189-195. \cr Hittner, J. B., May, K., & Silver, N.
C. (2003). A Monte Carlo evaluation of tests for comparing dependent
correlations. \emph{The Journal of general psychology}, 130(2), 149-168. \cr
Howell, D. C. (2013). \emph{Statistical methods for psychology} (8th ed.).
Belmont, CA: Wadsworth Cengage Learning.\cr Meng, X. L., Rosenthal, R., &
Rubin, D. B. (1992). Comparing correlated correlation coefficients.
\emph{Psychological Bulletin}, 111(1), 172-175. \cr Neill, J. J., & Dunn, O.
J. (1975). Equality of dependent correlation coefficients.
\emph{Biometrics}, 31(2), 531-543. \cr Olkin, I., & Finn, J. D. (1990).
Testing correlated correlations. \emph{Psychological Bulletin}, 108(2),
330-333. \cr Silver, N. C., Hittner, J. B., & May, K. (2004). Testing
dependent correlations with nonoverlapping variables: A Monte Carlo
simulation. \emph{The Journal of experimental education}, 73(1), 53-69. \cr
Steiger, J. H. (1980). Tests for comparing elements of a correlation matrix.
\emph{Psychological Bulletin}, 87(2), 245-251. \cr Wilcox, R. R. (2012).
\emph{Introduction to robust estimation and hypothesis testing}
}
\seealso{
Analysts are also encouraged to explore robust methods for
evaluation of correlation comparison hypotheses. For example, see work of R.
Wilcox (texts above and also
\emph{http://dornsife.usc.edu/labs/rwilcox/software/}
}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
