% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrtests.R
\name{test2r.mengz1}
\alias{test2r.mengz1}
\title{Test the difference between two dependent correlations with the the Meng z1
method}
\usage{
test2r.mengz1(ry.x1, ry.x2, rx1.x2, n)
}
\arguments{
\item{ry.x1}{y is the variable common to the two correlations.  It is
labeled y since the most common usage of this test is when a dependent
variable (y) is correlated with two different independent variables (x1, and
x2).  This argument \code{ry.x1} is the first of the two correlations of y
with the IV's.}

\item{ry.x2}{This argument \code{ry.x2} is the second of the two
correlations of y with the IV's.}

\item{rx1.x2}{The function and test require the Pearson correlation between
the two X's as well.}

\item{n}{Sample Size}
}
\value{
\item{z }{The test statistic value, a 'z'.} \item{pvalue
}{one-tailed probability of the 'z' test statistic.}
}
\description{
Differences in Pearson correlations are tested with the Meng's Z1
modification of Dunn's method.  The test is appropriate when the
correlations are dependent.  More specifically r(yx1) is tested versus
r(yx2)in one sample of cases.  The function requires the three Pearson
product moment correlations between three variables called y, x1 and x2 in
the notation here.  At present, the function only performs a one-tailed
test.
}
\details{
The Meng, et al., 1992 method uses the Fisher's Z transformation of the
Pearson correlation coefficients and produces a standard normal deviate.
}
\section{Related Functions}{
 \code{test2r.mengz1} is a member of a set of
functions that provide tests of differences between independent and
dependent correlations.  The functions were inspired by the \code{paired.r}
function in the \strong{psych} package and some of the code is modeled on code
from that function. See:
\itemize{
\item
\code{\link[bcdstats:test2r.t2]{test2r.t2}}, Test two dependent correlations
with the the T2 method: r(yx1) vs r(yx2)
\item
\code{\link[bcdstats:test2r.mengz1]{test2r.mengz1}}, the present function
\item
\code{\link[bcdstats:test2r.steigerz1]{test2r.steigerz1}}, Test the difference between
two dependent correlations with the the Steiger z1 method: r(yx1) vs r(yx2) in one
sample of cases.
\item
\code{\link[bcdstats:test2r.steigerz2]{test2r.steigerz2}}, Test the difference between
two dependent correlations with the the Steiger z2 method: r(jk) vs r(hm) in one
sample of cases.
\item
\code{\link[bcdstats:test2r.ind]{test2r.ind}} Test two r(xy) from
Independent Groups
}
}

\examples{

test2r.mengz1(.6,.31,.73,75)
test2r.mengz1(.45,.03,.65,100)
test2r.mengz1(.45,.03,.15,35)

}
\references{
Cheung, M. W. L., & Chan, W. (2004). Testing dependent
correlation coefficients via structural equation modeling.
\emph{Organizational Research Methods}, 7(2), 206-223. \cr Dunn, O. J., &
Clark, V. (1971). Comparison of tests of the equality of dependent
correlation coefficients. \emph{Journal of the American Statistical
Association}, 66(336), 904-908. \cr Hays, W. L. (1994). \emph{Statistics}
(5th ed.). Fort Worth: Harcourt College Publishers.\cr Hendrickson, G. F.,
Stanley, J. C., & Hills, J. R. (1970). Olkin's new formula for significance
of r13 vs. r23 compared with Hotelling's method. \emph{American Educational
Research Journal}, 7(2), 189-195. \cr Hittner, J. B., May, K., & Silver, N.
C. (2003). A Monte Carlo evaluation of tests for comparing dependent
correlations. \emph{The Journal of general psychology}, 130(2), 149-168. \cr
Howell, D. C. (2013). \emph{Statistical methods for psychology} (8th ed.).
Belmont, CA: Wadsworth Cengage Learning.\cr Meng, X. L., Rosenthal, R., &
Rubin, D. B. (1992). Comparing correlated correlation coefficients.
\emph{Psychological Bulletin}, 111(1), 172-175. \cr Neill, J. J., & Dunn, O.
J. (1975). Equality of dependent correlation coefficients.
\emph{Biometrics}, 31(2), 531-543. \cr Olkin, I., & Finn, J. D. (1990).
Testing correlated correlations. \emph{Psychological Bulletin}, 108(2),
330-333. \cr Silver, N. C., Hittner, J. B., & May, K. (2004). Testing
dependent correlations with nonoverlapping variables: A Monte Carlo
simulation. \emph{The Journal of experimental education}, 73(1), 53-69. \cr
Steiger, J. H. (1980). Tests for comparing elements of a correlation matrix.
\emph{Psychological Bulletin}, 87(2), 245-251. \cr Wilcox, R. R. (2012).
\emph{Introduction to robust estimation and hypothesis testing}
}
\seealso{
Analysts are also encouraged to explore robust methods for
evaluation of correlation comparison hypotheses. For example, see work of R.
Wilcox (texts above and also
\emph{http://dornsife.usc.edu/labs/rwilcox/software/}
}
\author{
Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
