% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdstats-package.R
\docType{package}
\name{bcdstats-package}
\alias{bcdstats-package}
\alias{bcdstats}
\title{A collection of functions for the APSY510/511 classes of B. Dudek}
\description{
Four Components to the package:\cr Graphing functions for univariate data
and descriptive summaries.\cr A set of functions to test the difference
between two pearson product-moment correlation coefficients.\cr Shiny apps
for visualization in statistics instruction, including interactive graphical
methods for probability distribution calculation/visualization.\cr A function to provide
supplemental information for \code{lm} regression models.
}
\details{
\tabular{ll}{ Package: \tab bcdstats\cr Type: \tab Package\cr Version: \tab
<<<<<<< HEAD
0.0.0.9005\cr Date: \tab 2022-08-15\cr License: \tab GPL-3\cr LazyLoad: \tab
=======
0.0.0.9004\cr Date: \tab 2021-08-10\cr License: \tab GPL-3\cr LazyLoad: \tab
>>>>>>> origin/master
yes\cr }
The bcdstats package provides graphical functions for univariate
data display, graphical displays for several probability distributions, and
significance test for evaluating differences in Pearson Product-Moment
Correlation coefficients.  The package is intended for usage by students in
the APSY510/511 classes.
}
\section{List of Functions and Data Frames in bcdstats}{
 See the
\code{Index} link at the bottom of this page.
}

\references{
See class documentation provided elsewhere.
}
\author{
Maintainer: Bruce Dudek \href{mailto:bruce.dudek@albany.edu}{bruce.dudek@albany.edu}
}
